requirements.txt
mcp
requests
beautifulsoup4
openai
python-dotenv
jose
